unify keyguard request types for client and server in one file
