+++ client/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@nimiq/keyguard-client",
+  "name": "test-keyguard-client",
   "version": "0.0.20",
-  "description": "Nimiq Keyguard client library",
+  "description": "Nimiq Keyguard client library dev test",
   "main": "dist/KeyguardClient.common.js",
   "module": "dist/KeyguardClient.es.js",
   "types": "types/index.d.ts",
