+++ client/src/main.ts
@@ -1,3 +1,3 @@
 export * from './KeyguardClient';
 export * from './RequestBehavior';
-export * from './RequestTypes';
+export * from './KeyguardCommand';
