+++ client/tslint.json
@@ -15,6 +15,7 @@
     "ordered-imports": false,
     "object-literal-sort-keys": false,
     "no-console": false,
+    "no-namespace": false,
     "max-classes-per-file": [3],
     "curly": [true, "ignore-same-line"],
     "variable-name": [true, "ban-keywords", "check-format", "allow-leading-underscore"]
