+++ src/request/export-words/ExportWords.js
@@ -9,7 +9,7 @@ class ExportWords extends Nimiq.Observable {
      * if a complete page is missing it will be created.
      * However these pages wil be the default pages which usually don't match the applications requirements.
      * Refer to the corresponsing _build(Privcy | RecoveryWords | ValidateWords) to see the general Structure.
-     * @param {ParsedSimpleRequest} request
+     * @param {KeyguardRequests.ParsedSimpleRequest} request
      * @param {Function} resolve
      * @param {Function} reject
      */
