+++ src/request/sign-transaction/LayoutStandard.js
@@ -4,7 +4,7 @@
 class LayoutStandard extends BaseLayout { // eslint-disable-line no-unused-vars
     /**
      * @param {?HTMLElement} $el
-     * @param {ParsedSignTransactionRequest} request
+     * @param {KeyguardRequests.ParsedSignTransactionRequest} request
      * @param {Function} resolve
      * @param {Function} reject
      */
